requirements.txt
django==1.11.4
psycopg2==2.7.3
django-autoslug==1.9.3
django-flatblocks==0.9.4
django-jquery==3.1.0
django-tinymce==2.4.0
Pillow==3.3.1
django-reversion==2.0.10